feat: merge sprint into calendar
